Cleaning up folder
I had a lot of stuff going on in this folder that wasn't meaningful or half finished and it was just messy in my opinion. It was nothing of significant value and wanted to start from scratch.
After deleting everything I added a simple RaylibStarter folder which when copied it will have all of the stuff needed to launch a Raylib window so no need to go through the trouble of getting it up and going again.