Game of Life logic finished
Everything regarding the actual logic for the game of life works. When ran the window comes up and using spacebar you can press through each iteration. Would like to add a timer so the user doesn't have to click through the iterations. Could also clean up the graphics a bit, but it works.