Create README.md
Added content to Readme file. It still needs a lot of work, but I wanted my brother to test it out on his machine and see if he could get it to work following the "How do I make these things work" section. He has no idea about coding, so he doesn't have any of the tools needed for it to work so he is a good test subject