Started Minesweeper
Copied RaylibBase (mainly sot the files are setup correctly to use, I don't really need all the comments anymore.

The goal is to have a desert background image, and then have the minesweeper game on top of that, and then have little grass squares that cover the minesweeper game (to start minesweeper all of the squares begin covered).

It doesn't look like a lot of progress was made, but was mainly messing around with images and cropping them and turning them into textures so they can be drawn to the screen.

Next time I will keep playing with images so it is correctly put on the screen or actually build the minesweeper game since those don't need to work together.