Added Visual
Created raylib visual. Currently instead of images is colored boxes, but shouldn't be that difficult to implement images into it. Bottom grid works, Top grid is currently displayed "over" bottom grid. Last thing added was receiving X and Y position of mouse when clicked. Now need to do the math to find out what square in the grid is clicked on and change it to be uncovered so bottom grid is shown.  After that have to create logic of uncovering multiple sqaures at a time (clicking on a square without a number reveals all squares connected that also dont have numbers)