Numbers added
I bought aseprite and created the numbers by myself. EW. But now when the grid is drawn it is drawn using the desired numbers instead of colored boxes. Also the functionality of using right click to uncover boxes does not work anymore because that function will be replaced with flagging mines.