Added Readme.txt
I think I am finished with this project at least for now. I learned a lot, but I definitely think there is a lot to improve on. This isn't a very big game and I shouldn't gotten it done a lot quicker, but I learned a lot as well. It would be cool to remake it some time in the future and see the difference in my starting project versus having done more. I think it is time to move on to something else though.