Simple Stuff
Didn't really do much. Creating the levels and having a character be in the level is intimidating. I am trying to watch some videos or learn how to do it, but I know its going be me just trying it before I can actually get a grasp of how it works.