Created Level Class
Created a level class that will read in .txt files where the levels can be manipulated. Also added a few functions to read, store, and print the level to make sure it works. Currently stuck on thinking how im going to draw the level with the character on it. But I think that might be a character thing? I dont know yet.. Probably create a player class and see where it goes from there next time.