Learning Wall collision
This is my first time making a game and making the wall collision is very intimidating. I got one movement (left) to work so it will stop at the wall, but it also clips into the wall a bit. I think I am going to add the others directions next and then try and add gravity. Probably fix the clipping later. Also the sprite size might fix the clipping problem on its own, we will see.