Added updatePlayer() and wall collision
updatePlayer() was added so there isn't a bunch of If(KeyPressed) lines in the main while loop. Now they are moved to a function for the player that will check all that and so it looks a little cleaner. Also the wall collison works. It probably isnt' right/modular but it works with how it is set up right now. There is a pixel or two of clipping on the bottom and right walls, but it works besides that. Up and Left dont clip anymore into the wall. I think next is gravity and then trying to add more floors.
Personal Note: Went home this past weekend which is why I didnt code Friday or Saturday. I know it is slow progress but I am also learning about this as I go so every small thing takes time now, but hopefully won't. Stay positive and keep coding. :)