Woking on clipping
I am thinking that adding multiple points to check on a side of the player should get rid of the clipping. For instance when falling, checking both the right and the left bottom corner of the player. I am having quite some trouble. I think my issue is when the distance to the ground is less than the step the player is about to take down. I am trying but it is frustrating