No movement while in shop
When the shop gui is open I will eventuallyw ant to use the same controls as movement to navigate the shop. I had to make it so the movement keys don't work while in the shop. Also dont want the player to be able to move around the map with the shop blocking the screen.