Added Character Class
Added Character class that Player and Monster will inherit from. Also added a level legend text file because I want everything to be in the text file and get created when the level is loaded, including the player and all the monsters/enemies. I feel like I am in over my head but one little step at a time.