Starting enum level values
It probably isnt necessary, but I wanted to give enums a try and eventually the level will be using a lot of random ASCII values so might as well give enums a try now. Currently it is in the level class, but I might make it its own class so that i can reference it in multiple places.