Added LevelValues class
Created an enum class LevelValues that can be referenced from anything that needs to see the level. I think it will make it easier to add new values into the text fields and to reference those values which actual names later. Makes it more readable. Also added moveLeft() which works almost as intended. A bit of clipping.