Added Shop()
Shop is now working as intended currently. Now just need to make it so player can purchase items in the shop. Probably time to merge this branch and make it the main branch since the character class is now added and at the same place now as the old main branch