First Item progress
I want to initially give the player single boots and have the item boots run the jump function. Currently trying to create the item for the boots and will eventually have the player buy it before the game starts. Haven't used much OOP before so it is a learning process.